add edite 26

+++ src/jvmMain/kotlin/common/component/ItemMenu.kt
@@ -1,25 +1,32 @@
 package common.component
 
-import androidx.compose.material.DropdownMenu
-import androidx.compose.material.DropdownMenuItem
-import androidx.compose.material.Text
+import androidx.compose.material.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Cancel
+import androidx.compose.material.icons.filled.ConfirmationNumber
+import androidx.compose.material.icons.filled.DeleteForever
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import features.sons_of_officers.domain.model.Person
+import styles.AppColors
 import styles.CairoTypography
 
+@OptIn(ExperimentalMaterialApi::class)
 @Composable
 fun ItemMenu(
     modifier: Modifier = Modifier,
     showMenu: MutableState<Boolean>,
     onEdit: () -> Unit,
     onRemove: () -> Unit,
+    showDialog: MutableState<Boolean>,
+//    onCancel: () -> Unit,
+    alertText:String
 ) {
-
     if (showMenu.value) {
         DropdownMenu(
             expanded = showMenu.value,
@@ -34,13 +41,50 @@ fun ItemMenu(
                 Text("تعديل", style = CairoTypography.h4)
             }
             DropdownMenuItem(onClick = {
-                // Handle delete action
-                onRemove()
-                showMenu.value = false
+                // Show AlertDialog
+                showDialog.value = true
             }) {
-
                 Text("مسح", style = CairoTypography.h4)
             }
         }
     }
+
+    if (showDialog.value) {
+        AlertDialog(
+            onDismissRequest = {
+                showDialog.value = false
+            },
+            title = {
+                Text(text = " ")
+            },
+            text = {
+                Text(alertText,textAlign = TextAlign.End,
+                    style = CairoTypography.h3)
+            },
+            confirmButton = {
+                GradientButton(
+                    text = "موافق",
+                    icon = Icons.Default.ConfirmationNumber,
+                    onClick = {
+                        onRemove()
+                        showDialog.value = false
+                    },
+                    colors = AppColors.blueGradient, cornerRadius = 30.dp
+                )
+            },
+            dismissButton = {
+
+                GradientButton(
+                    text = "الغاء",
+                    icon = Icons.Default.Cancel,
+                    onClick = {
+                        //                    onCancel()
+                        showDialog.value = false
+                    },
+                    colors = AppColors.RedGradient, cornerRadius = 30.dp
+                )
+            }
+        )
+    }
 }
+
